Split EntityWithMetadata into its own file
