+++ nixio/pycore/entity_with_sources.py
@@ -1,4 +1,12 @@
-from .entity import EntityWithMetadata
+# Copyright (c) 2016, German Neuroinformatics Node (G-Node)
+#
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted under the terms of the BSD License. See
+# LICENSE file in the root of the Project.
+
+from .entity_with_metadata import EntityWithMetadata
 from .source import Source
 from .util import util
